fix: move normalizers to encoder helpers
