+++ tests/unit_tests/encoder/time_series/test_timeseries_rnn.py
@@ -2,7 +2,7 @@ import unittest
 import numpy as np
 import torch
 from lightwood.encoder.time_series import TimeSeriesEncoder
-from lightwood.encoder.time_series.helpers.common import MinMaxNormalizer, CatNormalizer
+from lightwood.encoder.helpers import MinMaxNormalizer, CatNormalizer
 import pandas as pd
